Added Proxy for local testing
The Proxy is needed in order to make requests from localhost:3000 on localhost:3001 without any cross origin issues. This will only work locally, not on deployment. The idea is that your requests need to match in origin and because our server cant run on the same port our front end is running on locally, the server has to run on a different prot causing a different in origin, and this proxy handles the redirecting of requests